+++ src/app/views/admin/detalle-trabajador-admin/detalle-trabajador-admin.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ApiNa2Service } from 'src/app/services/api/api-na2.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-detalle-trabajador-admin',
@@ -29,15 +30,32 @@ export class DetalleTrabajadorAdminComponent {
     );
   }
   editarGerenteClick() {
-    this.apiService.actualizarTrabajador(this.trabajador).subscribe(
-      (data) => {
-        console.log(data);
-        this.modal.dismissAll();
-      },
-      (error) => {
-      }
-    );
+    Swal.fire({
+      icon: 'question',
+      title: "Editar Trabajador",
+      text: "¿Desea editar al Trabajador?",
+      showCancelButton: true,
+      confirmButtonColor: '#3CC3C8',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Editar',
+      cancelButtonText: 'Cancelar'
+    }).then(
+      (e) => {
+        if (e.isConfirmed) {
+          this.apiService.actualizarTrabajador(this.trabajador).subscribe(
+            (data) => {
+              Swal.fire("Exito", "Exito al editar", "success");
+              console.log(data);
+              this.modal.dismissAll();
+            },
+            (error) => {
+              Swal.fire("Error", "Erro al editar", "error");
+            }
+          );
+        }
+      });
   }
+
   validadoPDF() {
     this.apiService.obtenerRenuncia(this.trabajador.id).subscribe(
       (data) => {
@@ -70,16 +88,31 @@ export class DetalleTrabajadorAdminComponent {
   }
 
   eliminarGerente(idSurtidor: any) {
-    this.apiService.eliminarTrabajador(idSurtidor).subscribe(
-      (data) => {
+    Swal.fire({
+      icon: 'question',
+      title: "Eliminar Trabajador",
+      text: "¿Desea eliminar al trabajador?",
+      showCancelButton: true,
+      confirmButtonColor: '#3CC3C8',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar'
+    }).then(
+      (e) => {
+        if (e.isConfirmed) {
+          this.apiService.eliminarTrabajador(idSurtidor).subscribe(
+            (data) => {
 
-        this.router.navigate(['/admin/trabajadores']);
-      },
-      (error) => {
-        console.log(error);
-      }
-    );
+              this.router.navigate(['/admin/trabajadores']);
+            },
+            (error) => {
+              console.log(error);
+            }
+          );
+        }
+      });
   }
+
   listarinicidencias(id: any) {
     this.router.navigate(['/admin/incidencias/' + id], {
       queryParams: {
@@ -88,26 +121,42 @@ export class DetalleTrabajadorAdminComponent {
     });
   }
   generarPDF(idRenuncia) {
-    console.log(idRenuncia);
-    this.apiService.descargarRenuncia(idRenuncia).subscribe(
-      (data) => {
-        let downloadURL = window.URL.createObjectURL(data);
-        let link = document.createElement('a')
-        link.href = downloadURL;
-        link.download = "solicitudRenuncia.pdf"
-        link.click();
-      }, (error) => {
-        console.log(error);
-      }
-    );
+    Swal.fire({
+      icon: 'question',
+      title: "Descargar Archivo",
+      text: "¿Desea descargar el archivo?",
+      showCancelButton: true,
+      confirmButtonColor: '#3CC3C8',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Descargar',
+      cancelButtonText: 'Cancelar'
+    }).then(
+      (e) => {
+        if (e.isConfirmed) {
+          console.log(idRenuncia);
+          this.apiService.descargarRenuncia(idRenuncia).subscribe(
+            (data) => {
+              let downloadURL = window.URL.createObjectURL(data);
+              let link = document.createElement('a')
+              link.href = downloadURL;
+              link.download = "solicitudRenuncia.pdf"
+              link.click();
+            }, (error) => {
+              console.log(error);
+              Swal.fire("Error", "Error al descargar", "error");
+            }
+          );
+        }
+      });
   }
-  aceptar(id:any) {
+
+  aceptar(id: any) {
     console.log(id);
     this.apiService.aceptarRenuncia(id).subscribe(
-      (data)=>{
+      (data) => {
         console.log(data);
       }
-      
+
     );
     this.router.navigate(['/admin/trabajadores']);
     this.modal.dismissAll();
@@ -116,15 +165,31 @@ export class DetalleTrabajadorAdminComponent {
   no_aceptar() {
     this.validacion = true;
   }
-  editarRenunciaClick(){
-    this.apiService.actualizarRenuncia(this.trabajador.id, this.renuncia).subscribe(
-      (data) => {
-        console.log(data);
-        this.modal.dismissAll();
-      },
-      (error) => {
-      }
-    );
+  editarRenunciaClick() {
+    Swal.fire({
+      icon: 'question',
+      title: "Editar Renuncia",
+      text: "¿Desea editar renuncia",
+      showCancelButton: true,
+      confirmButtonColor: '#3CC3C8',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Editar',
+      cancelButtonText: 'Cancelar'
+    }).then(
+      (e) => {
+        if (e.isConfirmed) {
+          this.apiService.actualizarRenuncia(this.trabajador.id, this.renuncia).subscribe(
+            (data) => {
+              console.log(data);
+              this.modal.dismissAll();
+              Swal.fire("Exito", "Exito al editar", "success");
+            },
+            (error) => {
+              Swal.fire("Error", "Error al editar", "error");
+            }
+          );
+        }
+      });
   }
-  
+
 }
